BAU — Make some US/Canada postcode mapping constants static
Make some constants in the code that maps US ZIP codes to states
and Canadian postal codes to provinces or territories static, so
they’re only created once in the lifetime of the application.

The affected constants are all either regex patterns, unmodifiable
maps or strings, which are all immutable and thread-safe, so making
them static is safe.

Otherwise, each constant is created each time the class it’s in is
instantiated, which for some payment gateways can happen every time
an authorisation request is sent, even if the payment does not have
a US or Canadian address.