PP-6994 Use RefundService when calculating refundability
Instead of directly getting refunds from the database using the,
RefundDao use a new method in RefundService, which returns a list of
Refund objects.

This method will later be updated to retrieve refunds from ledger and
combine this set of refunds with the set retrieved from the connector
database so that we can handle calculating refundability when refunds
are expunged.