PP-6994 Don't use RefundDao to get refunds outside of RefundService
As we will now be retrieving refunds from both the connector database
and from ledger when calculating refundability, make it more explicit
in other places where we only retrieve non-expunged refunds from the
database by using a method in RefundService rather than directly calling
the RefundDao.

It is better practice that only the service uses the DAO, so removing
use of it from the resource is also a good thing to do.