PP-7229 Replace Joda-Time DateTime with java.time LocalDate
Stop using Joda-Time’s DateTime to represent the requested date to
capture a payment when a capture request is created to send to
Worldpay.

While the direct java.time equivalent of Joda-Time’s DateTime is
ZonedDateTime, use LocalDate instead because only the date
component is actually used. Retain the behaviour of determining the
current date based what the day is right now in the UTC time zone.

Stop setting a capture date altogether for authorisation requests
to Worldpay because it’s never used and doesn’t end up in the final
payload sent to Worldpay.