PP-7242 Use new CardExpiryDate type instead of strings
Use the new CardExpiryDate instead of strings in most places in
connector to represent card expiry dates.

This does not change how card expiry dates are represented in the
API (in either requests or responses), nor does it change how
card expiry dates are formatted when sent to payment gateways, nor
does it change how card expiry dates are stored in the database.

Due to the limitations of our current validation mechanism, there
is a change what happens when an invalid expiry date is supplied to
/v1/frontend/charges/{chargeId}/cards. Previously the validator
would catch it and return 400 response with a friendly error
message. Now it will fail when Jersey tries to deserialise it into
a CardExpiryDate object and a 400 response will be returned with
an exception message. This should be OK because only frontend calls
this endpoint and only for regular card payments (not Apple Pay or
Google Pay) and it has its own validation (tightened in PP-7080) so
should never send invalid expiry dates. The format required by
CardExpiryDate is actually stricter than the previous validation
that was in place for this endpoint but since frontend should only
send valid expiry dates this make no difference.

/v1/api/accounts/{accountId}/telephone-charges continues to
deserialise the expiry date to a string. It is only converted to a
CardExpiryDate object later, after the endpoint has performed its
own validation, which allows card expiry dates in the exact same
format as required by CardExpiryDate (and now it actually uses
CardExpiryDate’s code to check the format). Therefore the behaviour
of this endpoint when supplied with an invalid expiry date should
not have changed.

The mechanism by which 2-digit years are made into 4-digit years by
prefixing them with “20” when sending them in authorisation
requests to Worldpay and Smartpay remains but this logic is now
encapsulated in CardExpiryDate rather than in the Freemarker XML
templates.

Prior to this change, an expiry date associated with a payment was
converted to a string when retrieved from the database. Now it is
immediately converted to a CardExpiryDate object, which will fail
if the expiry date in the database is not in the required format. A
check of all the expiry dates stored in the production database a
couple of days ago confirmed that all are in the correct format and
we do not expect this to have changed since then due to the
existing validation we have in place in connector and frontend.

Expiry dates in payments retrieved from ledger remain as strings,
as is the convention in this area of code (in addition, we haven’t
verified that every expiry date in the ledger database matches the
required format). Expiry dates that are sent to ledger in events
are converted to strings prior to the event object being serialised
to JSON; again, this matches the existing conventions in this area
of code.