PP-7242 Add test to check invalid card expiry date still rejected
We no longer validate the card expiry date when processing card
details from a regular card payment, instead relying on the
instantiation of CardExpiryDate to fail if it’s not legit.

Add a test for this behaviour at the integration level to make up
for the fact we don’t have validation tests that cover this any
more.

Adding this test required changing some variables back from
CardExpiryDates to strings so that we can test with invalid values
(since CardExpiryDate can’t represent invalid values). However,
this is limited only to integration tests that construct JSON
payloads.