PP-6994 Do not re-request ledger refunds during refund
After making a network request to get the external refund-ability, we
were re-querying the refunds in the database to ensure there had been
no changes in the meantime.

When the change was made to also get refunds that only exist in ledger
for the refund-ability calculation, in order to allow us to expunge
refunds from connector, this was changed to also re-query ledger.

We shouldn't need to re-query ledger as the expunged refunds will be in
a terminal state - with the exception of being able to be updated by an
admin.

We should however only really care if the amount refundable has
decreased rather than increased. And in the case of an admin updating
the status of a historic refund from "error" -> "success", we would
expect the external refund-ability check with the payment provider to be
unsuccessful.

It could also be possible, depending on how long we wait before
expunging refunds in the future, that a refund is expunged before we
query the database a second time. If this is the case, we are fine to
continue using the record we obtained the first time without re-querying
ledger as it is not possible that the amount available to refund would
have decreased, even if the status of that refund had changed in the
meantime.