PP-6991: Split out DnsUtils
Split out DnsUtils into its two functions: an IP to domain matcher and reverse
DNS lookup.

Doing this will allow the reverse DNS lookup to be stubbed out in tests, thus
removing the need for an actual network to be present. We don't need tests for
the reverse DNS lookup as code to do this is from a cookbook and not part of
our business logic.