PP-7205 Use commons version of RestClientLoggingFilter
This has been moved to pay-java-commons to reduce duplication.

Also, ensure we are always putting the request id on the MDC with the
key "x_request_id". Previously it was added with the key "X-Request-Id"
for some of our tasks meaning it was logged with this key in some places
and in other places was not logged at all.