PP-7336: Update Stripe to use a list of allowed CIDRs for firewall
This change updates Stripe Gateway Config to pick up a list of allowed CIDRs instead of IP addresses because using CIDR helps to reduce a large number of IP addresses needed to be specified in the config file. It also updates IP Address Matcher to validate a forwarded IP address before processing it.