PP-7417: Rename WorldpayUpdate3dsFlexCredentialsRequest
The structure of Worldpay3dsFlexCredentialsRequest is exactly the same as the
request to validate 3DS Flex credentials in the up-and-coming
`/v1/api/accounts/1/worldpay/check-3ds-flex-config` url, so it makes sense to
rename the class. Note that in the future validation annotations will be added
to make sure the organisational_unit_id and issuer are of 24 hexadecimal
characters and the jwt_mac_key is in UUID format.