PP-7524 Refactor getGatewayAccountCredentialsAsAuthHeader
To help remove the coupling between our database and API layer remove
the need for the GatewayAccountEntity when constructing the auth header
to call a payment provider. It only needs a map of credentials. This
will help use to reuse our PaymentProvider interface for charges which
have been expunged from Connector.