PP-7559 If we need to send an email but the charge doesn't have one log a warning
We are logging an error when we need to send an email based on the
current gateway account email collection settings but the charge doesn't
have an email address on it. This appears to happen for Telephone
payments which can be created without an email address despite the
gateway accounts email collection settings. We might want to address
that in the validation for creating telephone payments but if it gets to
here we cannot do anything about it so log as a warning.