PP-5641: Build auth order in response to soft decline
In the event where Worldpay responds to an exemption request with a soft
decline, the auth request should be sent again with no <exemption> element.
This commit doesn't contain code to retry the auth request, but to ensure an
auth request without an <exemption> element can be created even for a gateway
account with exemption engine enabled.