bau: Refactor WorldpayPaymentProvider
Initialise WorldpayPaymentProvider with its required dependencies, rather than
creating them within its constructor. This will make it easier to test future
functionality. For example, an authorisation request will need to be retried if
worldpay soft declines an exemption request. In this case all we'd need to test
is that `WorldpayAuthoriseHandler.authorise` is called twice.

It also improves on what we had before, where we had to mock internal
dependencies which is bad practice and makes reading tests difficult.