PP-7590 Use Instant not ZonedDateTime for createdDate in Charge
Update the Charge class so that the createdDate field is an Instant
rather than a ZonedDateTime, which is line with ChargeEntity.

When creating a Charge from a LedgerTransaction, where we obtain
the createdDate by parsing a string, we now use Instant.parse(…)
rather than ZonedDateTime.parse(…). This accepts a narrower range
of formats but everything ledger gives us should be an ISO 8601
string, which is compatible with Instant.parse(…). Some tests have
been updated to not use ZonedDateTimes in the default time zone to
reflect this.

Also change DateTimeUtils.toUserFriendlyDate(…) to take an Instant
rather than a ZonedDateTime because all the callers have Instants
now.