BAU Replace YYYY with yyyy in DateTimeFormatter
`YYYY` will return the year based upon the week year based on the locale
setting. This can give suprising results near year end and in
circumstances when the locale is not as expected, eg. `en_US`
rather than `en_GB`.

`yyyy` will return the calendar year and is what we most likely want.

Intelij highlights this method and others in the
`AppleDecryptedPaymentData` as unused but they are magically called by
the template builder as part of
`WorldpayWalletAuthorisationHandler.buildWalletAuthoriseOrder` method.