PP-7643: Remove exemption3ds from WorldpayAuthorisationRequestSummary
Due to the difficulties in keeping the logging consistent for different
exemption engine scenarios (e.g. soft decline where the authorisation is
retried within the WorldpayPaymentProvider), we have decided to just not log
whether exemption3ds was requested. This is ok as we log the exemption3ds on
the charge anyway which implicitly tells us this information.