PP-5025: Help reduce key and cert creation errors
Creating the Apple private key and cert can be quite gnarly. The openssl
command that generates the private key, as well as the payment processing
certificate received from Apple, have multiple lines. We usually manually edit
these files to make them a single line which is what connector expects. We have
seen recently that in the process of doing this manually the lines are joined
with a space in between them. This commit helps with the fiddliness of creating
the private key and cert by removing any whitespaces in case they exist.

We have also seen that when storing an empty password in
https://github.com/alphagov/pay-infra/blob/master/cli/config/secrets.yml, for
example:
```
APPLE_PAY_PAYMENT_PROCESSING_CERTIFICATE_SECONDARY:       ""
```
the java fields (`secondaryPrivateKey`, `secondaryPublicCertificate`) that the
environment variable gets bound to might be an empty string rather than null.
The secondary Apple payment keycerts being optional, return nulls if the string
is blank.