PP-7622 Add test for new Apple Pay Certs
Two new tests encrypt the payload with either the primary or secondary
public keys as Apple would and then test that our ApplePayDecrypter can
decrypt the payload using the appropriate private key.

This was originally developed to help test the rotation of certs in an
environment but it seems to have merit in itself and will be useful in
future for testing new keypairs when we roll them.

The secondary cert and private key used in the tests are the old and now
revoked cert and key that we were using in test-12.