PP-8255 Send payer email address to Worldpay if enabled
If a Worldpay gateway account has the setting to send the payer
email address to the gateway enabled, and there is an email address
for the payer (either entered directly by the user or passed to us
by the service when they created the payment), send it to Worldpay
in the authorisation request for a regular card payment.

The email address is sent in the XML authorisation request to
Worldpay in an element called <shopperEmailAddress>, which is in
turn inside an element called <shopper> (if 3D Secure is enabled,
we already send a <shopper> element).

The <shopper> element appears after the <paymentDetails> element.
If 3D Secure is disabled, the email address is sent like this:

<shopper>
    <shopperEmailAddress>payer@email.example</shopperEmailAddress>
</shopper>

If 3D Secure is enabled, the email address is sent before the
<browser> element (this matches Worldpay’s examples in their
documentation and worked when tested):

<shopper>
    <shopperEmailAddress>payer@email.example</shopperEmailAddress>
    <browser>
        <acceptHeader>*/*</acceptHeader>
        <userAgentHeader>Mozilla/5.0</userAgentHeader>
    </browser>
</shopper>

There are no changes if either the setting to send the payer email
to the gateway is not enabled, or if there is no payer email
address available.

The email address is only sent for regular card payments. It is not
sent for Apple Pay or Google Pay payments.

Worldpay have a setting to send payment confirmation emails to
paying users if an email address is supplied. To avoid confusing
paying users, this setting should not be enabled.