PP-8446: Allow gateway accounts to be filtered on service ID
The `service_id` column on the `gateway_accounts` table has a standard
`btree` index.

Add support for filtering the `/v1/frontend/accounts` route for accounts
that belong to a particular list of service identifiers with the
`serviceIds` parameter.

There are a number of changes that could be made to the param -> filter
methodology here but this is implemented to follow the pattern already
established.