PP-8628 Send capture success details for forced capture event
When we receive a capture success notification, but the charge is in an
unexpected state, we ignore the current state and update the status to
captured and emit an event to ledger to do this.

However, this event did not include the capture details such as the
captured date. We were also not storing the gateway_event_date for the
event in connector so the capture event was derived from the updated
date for the event rather than the date from the notification.

Fix both this by storing the gateway_event_date and attaching
the same event details we attach for the standard CAPTURE_CONFIRMED
event to the STATUS_CORRECTED_TO_CAPTURED_TO_MATCH_GATEWAY_STATUS event.