PP-8945 Store fees and send event for failed payments
When processing Stripe failed payments, after transferring the
applicable amount for Stripe fees from the Stripe connect account, store
entries in the fees table for each of the fee types that were applied.
Send a FEE_INCURRED event to ledger containing the fee breakdown.

Always expect that the charge exists in the connector database, and
throw an exception if it has been expunged. We will separately add a
check that will stop Stripe charges from being expunged if they have not
had the applicable fees taken.