PP-8945 Fix pact tests
Pact tests were failing because ledger was expecting a positive integer
for the net_amount, but was receiving a negative value as the charge we
were using to build the event details for the captured events did not
have the CAPTURED status.

This change was required as we now return a negative net_amount when the
charge is not captured to allow for reporting on fees for failed
Stripe payments.