PP-8945 Handle Stripe charge not existing
For some payments that are fail 3DS, no charge is created for the
payment intent in Stripe. This means that we cannot determine from the
Stripe API whether or not the payment has been through 3DS, and so
whether to apply a 3DS fee.

Implement a (possibly) temporary solution, where if there is no Stripe
charge, look at the charge events in connector to determine whether 3DS
was completed. We are only able to determine that 3DS was used if there
is an AUTHORISATION_REJECTED event or AUTHORISATION_SUCCESS event. This
means we cannot account for payments where 3DS is initiated but not
completed, which we may still be charged a fee for.