PP-8951 Add reason to metadata for Stripe transfers
Add a "reason" field to the metadata we attach to a Stripe transfer when
it is created. This will allow us to later determine the purpose of the
transfer, as we need to be able to differentiate between transfers made
for refunds and transfers made to collect fees for failed payments when
reconciling payouts.

We also attach a reason to transfers made for payments when they are
captured, but we will not use this for anything other than information
purposes when viewing in the Stripe dashboard.

Co-authored-by: Nathaniel Steers <nathaniel.steers@digital.cabinet-office.gov.uk>