PP-9265 Force shutdown payout queue receiver on shutdown
It is fine to kill threads running tasks to process payout
reconciliation immediately, as this just sends events to ledger for the
associated payments and is retry-able with no adverse effects.

It is desirable to do this, as the thread running the task can be
waiting for up to 20s for messages on the SQS queue to become available
as we use long polling, and so block the connector shutdown.

For most of the day there are no payouts to process, so the task will
usually be waiting for the entire 20s for messages to become available.